Build6: added shopping cart logic fixes, some jQuery div hiding for email building screen realestate, and other general fixes.
Real work going on is the finalisation of the shopping cart system
Checkout function to place orders and clear the cart.
Edit items in cart page with delete button and multiple edit of qty -- only if time persists.

Documentation and some Administrator walkthrough videos (short in length)

Testing, testing and more testing on all Platforms / devices.

Yay were almost ready to hand this off to the next team to finish the MYOB functionality and the other areas we did not get time to finish off.